% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{color_palettes}

\alias{color_palettes}

\title{Color palettes for reproducing original graphs}

\description{
A dataset containing all color palettes used in original graphs
}

\usage{
data(color_palettes)
}

\format{
list with each member representing one color palette.
}

\keyword{datasets}


\docType{data}
