% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customFunctions.R
\name{filterForVoom}

\alias{filterForVoom}

\title{Filtering a count matrix to remove low read genes}

\description{
This function takes a count matrix and removes genes that don't reach a minimum namer of reads in a certain number of samples.
}

\usage{
filterForVoom(m, minSamples = ncol(m)/10, minCpm = 0.25)
}

\arguments{
\item{m}{raw count matrix (genes are rows, columns are samples)}

\item{minSamples}{numeric; minimum number of samples in which CPM threshold has to be reached; defaults to a tenth of available samples}

\item{minCpm}{numeric; the minimum counts per million (CPM) that has to be reached}
}

\value{
filtered count matrix
}

\author{
Dorothee Nickles
}
