% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customFunctions.R
\name{filterNvoom}

\alias{filterNvoom}

\title{Filter and normalize a count matrix}

\description{
This wrapper function combines filtering out genes with low reads in a number of samples (recommended for limma:voom) with normalization
}

\usage{
filterNvoom(m, minSamples = ncol(m)/10, minCpm = 0.25, DESeq = FALSE)
}

\arguments{
\item{m}{raw count matrix (genes are rows, columns are samples)}

\item{minSamples}{numeric; minimum number of samples in which CPM threshold has to be reached; defaults to a tenth of available samples}

\item{minCpm}{numeric; the minimum counts per million (CPM) that has to be reached}

\item{DESeq}{logical; if set to FALSE (default), edgeR normFactors will used for normalization; otherwise, DEseq sizeFactors will be used}
}

\value{
filtered and normalized count matrix
}

\author{
Dorothee Nickles
}
