% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fmone}

\alias{fmone}

\title{Targeted mutation data for the participants of the IMvigor210 trial}

\description{
A dataset containing the mutation calls based on FMOne panel and basic sample annotation for
participants of the IMvigor210 trial (NCT02108652/NCT02951767).
}

\usage{
data(fmone)
}

\format{
An NChannelSet containing sample annotations as well as all known and likely variants,
except for chromosomal rearrangements, as reported by Foundation Medicine, Inc.
Each variant type is stored in a separate channel.
\describe{
  \item{amplifications}{copy number alterations of genes that are recurrently amplified in cancer}
  \item{deletions}{copy number alterations of genes that are recurrently deleted in cancer}
  \item{gains}{non-focal amplifications (> 20MB) recurrently amplified in cancer}
  \item{known_short}{short variants (< 49bp long) that are known to have deleterious effects in cancer}
  \item{likely_short}{short variants (< 49bp long) that are likely to have deleterious effects in cancer}
  \item{ANONPT_ID}{Identifier, mock patient ID}
  \item{Best Confirmed Overall Response}{response to treament, by RECISTv1.1 and independent radiology facility; four levels: progressive disease (PD), stable disease (SD), partial response (PR) and complete response (CR)}
  \item{binaryResponse}{binarized response, based on 'Best Confirmed Overall Response', grouping PD/SD patients into one group, CR/PR patients into another}
  \item{IC Level}{level of IHC-assessed PD-L1 staining on immune cells (IC), with IC0 meaning <1 percent, IC1 >= 1 percent but <5 percent, IC2+ >= 5 percent immune cells staining for PD-L1}
  \item{TC Level}{level of IHC-assessed PD-L1 staining on tumor cells (TC), with TC0 meaning <1 percent, TC1 >= 1 percent but <5 percent, TC2+ >= 5 percent tumor cells staining for PD-L1}
  \item{FMOne mutation burden per MB}{tumor mutation burden, i.e. number of mutations per MB; as estimated by FMOne panel}
}
}

\keyword{datasets}


\docType{data}
