% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customFunctions.R
\name{getHRandCIfromCoxph}

\alias{getHRandCIfromCoxph}

\title{Extract hazard ratio and confidence intervals from a coxph object}

\description{
Convenience function to extract hazard ratio and confidence intervals from a coxph object,
as generated by a call to coxph(), comparing survival times between two groups
}

\usage{
getHRandCIfromCoxph(coxphData)
}

\arguments{
\item{coxphData}{coxph object; coxphData <- coxph(Surv(TTE, Cens) ~ group, data=df)}
}

\author{
Dorothee Nickles
}
