% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customFunctions.R
\name{human2mouse}

\alias{human2mouse}

\title{Convert human gene symbols to mouse gene symbols using biomaRt.
Our gene signatures consist of human gene symbols, so in order to score mouse gene expression
samples for gene signatures, we first need to convert the human gene symbols to mouse gene symbols.}

\description{
Convert human gene symbols to mouse gene symbols using biomaRt.
Our gene signatures consist of human gene symbols, so in order to score mouse gene expression
samples for gene signatures, we first need to convert the human gene symbols to mouse gene symbols.
}

\usage{
human2mouse(x)
}

\arguments{
\item{x}{character vector for human gene symbols}
}

\value{
matrix; N by 2 matrix where the first column is the input human gene symbols, and the second column is the corresponding mouse gene symbols. One human gene can map to multiple mouse genes, and also multiple human genes can map to one mouse gene.
}

\author{
Yasin Senbabaoglu
}
