% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customFunctions.R
\name{matrix4signature}

\alias{matrix4signature}

\title{Auxiliary function that prepares a gene by sample matrix for the computation of signature scores.
If a human gene has multiple orthologs in mouse, then the mouse gene with maximum MAD (median absolute deviation) across the samples is used.}

\description{
Auxiliary function that prepares a gene by sample matrix for the computation of signature scores.
If a human gene has multiple orthologs in mouse, then the mouse gene with maximum MAD (median absolute deviation) across the samples is used.
}

\usage{
matrix4signature(
  x,
  genelist,
  gprofile,
  conversionDB,
  lowExpressionCutoff = NULL
)
}

\arguments{
\item{x}{gene expression matrix; normalized (CPM, TPM, RPKM, FPKM, etc.) but not standardized across samples; samples in columns, genes in rows}

\item{genelist}{named list object where each element is a character vector and comprises of a gene signature}

\item{gprofile}{character vector; gene symbols for all genes in x (in the same order as in x)}

\item{conversionDB}{N by 2 matrix; output of human2mouse(); conversion database showing the mapping from human to mouse gene symbols}

\item{lowExpressionCutoff}{numeric; low expression cutoff for the genes in a signature. If this is defined, genes with maximum expression value (maximum across samples) less than this cutoff are removed from the signature.}
}

\value{
matrix; gene expression matrix with genes in rows and samples in columns
}

\author{
Yasin Senbabaoglu
}
