% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customFunctions.R
\name{mySimpleVoom}

\alias{mySimpleVoom}

\title{Normalize a count matrix using limma::voom}

\description{
This function takes a count matrix, calculates either DESeq sizeFactors or EdgeR normFactors and normalizes counts using limma::voom
}

\usage{
mySimpleVoom(m, DESeq = FALSE)
}

\arguments{
\item{m}{raw count matrix (genes are rows, columns are samples)}

\item{DESeq}{logical; if set to FALSE (default), edgeR normFactors will used for normalization; otherwise, DEseq sizeFactors will be used}
}

\value{
normalized count matrix
}

\author{
Dorothee Nickles
}
