% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customFunctions.R
\name{pLevel}

\alias{pLevel}

\title{Convert numeric p-values into significance levels}

\description{
This convenience function converts a numeric p-value into "***" annotation, indicating significance level
}

\usage{
pLevel(pval)
}

\arguments{
\item{pval}{numeric vector; p-value(s) to be converted}
}

\value{
character vector; "***" = P < 0.001, "**" = P < 0.01, "*" = P < 0.05, "." < P < 0.1, "n.s." = not significant
}

\author{
Dorothee Nickles
}
