% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customFunctions.R
\name{plotSurvival2}

\alias{plotSurvival2}

\title{Custom version of Kaplan Meier plot}

\description{
Custom display of survival data, including plotting lines for median survival and adding number at risk
}

\usage{
plotSurvival2(
  survFit,
  survDiff,
  diff.factor,
  main,
  cols,
  xLab = "PFS (months)",
  yLab = "Probability of survival",
  ltypes = 1,
  plotMedian = FALSE,
  pval = c("logrank", "coxph", "none"),
  coxphData,
  legPos = "topright",
  mar = c(12, 9, 3, 2),
  plot.nrisk = FALSE,
  nrisk.interval = 2,
  timemark = TRUE,
  lwd = 2,
  cexMedian = 0.75,
  cexLegend = 0.8,
  ...
)
}

\arguments{
\item{survFit}{survfit object; survfit(Surv(TTE, Cens) ~ group, data=df)}

\item{survDiff}{surfdiff object; survdiff(Surv(TTE, Cens) ~ group, data=df)}

\item{diff.factor}{factor, the grouping factor used for survfit call}

\item{main}{character; main title for plot}

\item{cols}{character vector; colors, default to darkgreen, darkmagenta, cyan4, darkorange, darkred}

\item{xLab}{character; x-axis label for plot}

\item{yLab}{character; y-axis label for plot}

\item{ltypes}{numeric vector; line types (lty) for plot}

\item{plotMedian}{logical; shall the median survival for each group be plotted}

\item{pval}{type of p-value computed, either "logrank" or "coxph" (the latter
can only be computed on two groups)}

\item{coxphData}{coxph object; coxphData <- coxph(Surv(TTE, Cens) ~ group, data=df)}

\item{legPos}{legend position for legend()}

\item{mar}{plotting margins}

\item{plot.nrisk}{logical; shall number of samples for each group and time point
be plotted underneath graph}

\item{nrisk.interval}{numeric; spacing of time intervals for which number of samples
is given; defaults to 2}

\item{timemark}{logical; if set to TRUE (default), censoring marks are added to survival curves}

\item{lwd}{line width for survival curves}

\item{cexMedian}{numeric; relative size of font for median survival times}

\item{cexLegend}{numeric; relative size of legend font}

\item{...}{other arguments passed to plot.survfit}
}

\author{
Yuanyuan Xiao, Dorothee Nickles
}
