% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customFunctions.R
\name{any_mutation}
\alias{any_mutation}
\title{Combine data from all mutation matrices}
\usage{
any_mutation(ncs)
}
\arguments{
\item{ncs}{The \code{\linkS4class{NChannelSet}} object.}
}
\value{
A logical matrix with row and column names matching the input object,
suitable for direct addition as additional assay data.

Assay data matrices are assumed to contain a non-empty string if a mutation
was found, and a zero-character string if nothing was found.
}
\description{
Given an \code{\linkS4class{NChannelSet}} containing FMOne mutation data, check all assay data mutation matrices.
If a non-empty string is found in one or more, return TRUE.
}
\author{
Richard Bourgon
}
