% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cds}

\alias{cds}

\title{Count data for the transcriptomes of the participants of the IMvigor210 trial}

\description{
A dataset containing the transcriptomes and sample annotation for the majority of
participants of the IMvigor210 trial (NCT02108652/NCT02951767).
}

\usage{
data(cds)
}

\format{
A DESeqDataSet of raw counts for all genes as well as basic feature and
all sample annotations reported in the manuscript:
\describe{
  \item{ANONPT_ID}{Identifier, mock patient ID}
  \item{Best Confirmed Overall Response}{response to treament, by RECISTv1.1 and independent radiology facility; four levels: progressive disease (PD), stable disease (SD), partial response (PR) and complete response (CR)}
  \item{binaryResponse}{binarized response, based on 'Best Confirmed Overall Response', grouping PD/SD patients into one group, CR/PR patients into another}
  \item{IC Level}{level of IHC-assessed PD-L1 staining on immune cells (IC), with IC0 meaning <1 percent, IC1 >= 1 percent but <5 percent, IC2+ >= 5 percent immune cells staining for PD-L1; sample-level}
  \item{os}{overall survival in months}
  \item{censOS}{censoring information for surival data, with 0=alive and 1=dead}
  \item{Lund}{cluster membership for Lund molecular subtypes}
  \item{Lund2}{Lund molecular subtypes}
  \item{TCGA Subtype}{TCGA molecular subtypes}
  \item{Enrollment IC}{Representative IC level of a patient that was used during trial enrollment; patient-level}
  \item{Immune phenotype}{tumor immune phenotype, based on IHC assessment; one of the three: inflamed - robust CD8+ T-cell infiltration and PD-L1 expression, excluded - T cells accumulating in the extracellular matrix–rich stroma, desert - paucity of infiltrating lymphocytes within the tumour or surrounding stroma}
  \item{Sex}{sex, either M (male) or F (female)}
  \item{Race}{race}
  \item{Intravesical BCG administered}{(Y)es or (N)o}
  \item{Baseline ECOG Score}{ranging from 0-2; explanation at http://ecog-acrin.org/resources/ecog-performance-status}
  \item{Tobacco Use History}{either NEVER, PREVIOUS, or CURRENT}
  \item{Met Disease Status}{describing whether metastases affect lymph nodes only (LN Only), affect also visceral tissue (Visceral) or are present in the liver (Liver)}
  \item{TC Level}{level of IHC-assessed PD-L1 staining on tumor cells (TC), with TC0 meaning <1 percent, TC1 >= 1 percent but <5 percent, TC2+ >= 5 percent tumor cells staining for PD-L1; sample-level}
  \item{FMOne mutation burden per MB}{tumor mutation burden, i.e. number of mutations per MB; as estimated by FMOne panel}
  \item{Sample age}{three groups - more than two years: sample collected more than 2 years, 1-2 years: sample collected within 1-2 years before treatment start, (less than) 1 year: sample collected most recently}
  \item{Tissue}{tissue origin of sample}
  \item{Received platinum}{Has the patient ever received platinum treatment? (Y)es/(N)o}
  \item{Sample collected pre-platinum}{If patient has received platinum, was sample collected before treatment? (Y)es/(N)o}
}
}

\source{
EGAS00001002556
}

\keyword{datasets}


\docType{data}
