% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{color_palettes}
\alias{color_palettes}
\title{Color palettes for reproducing original graphs}
\format{list with each member representing one color palette.}
\usage{
data(color_palettes)
}
\description{
A dataset containing all color palettes used in original graphs
}
\keyword{datasets}
