% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customFunctions.R
\name{limitRange}
\alias{limitRange}
\title{Cap values in a numeric matrix at a specific value}
\usage{
limitRange(m, max.z = 3)
}
\arguments{
\item{m}{numeric matrix}

\item{max.z}{numeric vector of length 1; maximum/minimum number at which data will be truncated}
}
\value{
capped numeric matrix
}
\description{
This convenience function sets an absolute maximum a value cannot exceed in a numeric count matrix
}
\author{
Thomas Sandmann
}
