% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customFunctions.R
\name{score.signatures.DB.mouse}
\alias{score.signatures.DB.mouse}
\title{Score mouse gene expression samples for signatures.}
\usage{
score.signatures.DB.mouse(x, glist, gprofile, conversionDB,
  log.inside = FALSE, print.sig.names = TRUE, lowExpressionCutoff = NULL)
}
\arguments{
\item{x}{gene expression matrix; normalized (CPM, TPM, RPKM, FPKM, etc.) but not standardized across samples; samples in columns, genes in rows}

\item{glist}{named list object where each element is a character vector and comprises of a gene signature}

\item{gprofile}{character vector; gene symbols for all genes in x (in the same order as in x)}

\item{conversionDB}{N by 2 matrix; output of human2mouse(); conversion database showing the mapping from human to mouse gene symbols}

\item{log.inside}{boolean; should x be log-transformed inside the function?, log-transformation is recommended, can be done inside or outside the function}

\item{print.sig.names}{boolean; should the signature names be printed on the screen as they are being scored?}

\item{lowExpressionCutoff}{numeric; low expression cutoff for the genes in a signature. If this is defined, genes with maximum expression value (maximum across samples) less than this cutoff are removed from the signature.}
}
\value{
matrix; scores for signatures; signatures in rows, samples in columns
}
\description{
Score mouse gene expression samples for signatures.
}
\author{
Yasin Senbabaoglu
}
